+++ src/builder/stage.interface.ts
@@ -1,6 +1,7 @@
 export interface IStage {
   $match?: any;
-  $group?: any;
   $sort?: any;
   $limit?: number;
+  $skip?: number;
+  $group?: any;
 }
